requirements.txt
torch==2.0.1
deepspeed==0.11.1
tokenizers==0.14.1
huggingface-hub==0.23.0
jsonlines==4.0.0
accelerate==0.32.1
matplotlib==3.8.0
nltk==3.8.1
numpy==1.25.2
openai==0.28.0
packaging==24.0
pandas==2.2.2
peft==0.10.0
sentencepiece==0.1.99
Pillow==10.3.0
requests==2.32.3
shortuuid==1.0.13
spacy==3.7.2
timm==0.9.10
tqdm==4.66.1
transformers==4.35.0
wandb==0.15.11
opencv-python==4.9.0.80
datasets==2.21.0
tensorboardX==2.6.2.2